+++ hololinked/server/zmq_message_brokers.py
@@ -2,6 +2,7 @@ import builtins
 import os
 import threading
 import time
+import warnings
 import zmq
 import zmq.asyncio
 import asyncio
@@ -2188,7 +2189,21 @@ class EventPublisher(BaseZMQServer, BaseSyncZMQ):
             raise AttributeError(f"event {event._name} already found in list of events, please use another name.")
         self.event_ids.add(event._unique_identifier)
         self.events.add(event) 
-        
+    
+    def unregister(self, event : "EventDispatcher") -> None:
+        """
+        unregister event with a specific (unique) name
+
+        Parameters
+        ----------
+        event: ``Event``
+            ``Event`` object that needs to be unregistered. 
+        """
+        if event in self.events:
+            self.events.remove(event)
+            self.event_ids.remove(event._unique_identifier)
+        else:
+            warnings.warn(f"event {event._name} not found in list of events, please use another name.", UserWarning)
                
     def publish(self, unique_identifier : bytes, data : typing.Any, *, zmq_clients : bool = True, 
                         http_clients : bool = True, serialize : bool = True) -> None: 
@@ -2215,10 +2230,15 @@ class EventPublisher(BaseZMQServer, BaseSyncZMQ):
                     return
                 if zmq_clients:
                     # TODO - event id should not any longer be unique
-                    self.socket.send_multipart([unique_identifier, self.zmq_serializer.dumps(data)])
+                    self.socket.send_multipart([b'zmq-' + unique_identifier, self.zmq_serializer.dumps(data)])
                 if http_clients:
                     self.socket.send_multipart([unique_identifier, self.http_serializer.dumps(data)])
-            else:
+            elif not isinstance(self.zmq_serializer , JSONSerializer):
+                if zmq_clients:
+                    self.socket.send_multipart([b'zmq-' + unique_identifier, data])
+                if http_clients:
+                    self.socket.send_multipart([unique_identifier, data])
+            else: 
                 self.socket.send_multipart([unique_identifier, data])
         else:
             raise AttributeError("event name {} not yet registered with socket {}".format(unique_identifier, self.socket_address))