+++ CHANGELOG.md
@@ -9,9 +9,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ✓ means ready to try
 
+New:
 - cookie auth & its specification in TD (cookie auth branch)
-- image event handlers (develop branch) for streaming live video as JPEG and PNG ✓
-- pydantic support for property models (develop branch) ✓
+- adding custom handlers for each property, action and event to override default behaviour
+- pydantic support for property models 
+
+Bug Fixes:
+- composed sub`Thing`s exposed with correct URL path ✓ 
+
+## [v0.2.7] - 2024-10-22
+
+- HTTP SSE would previously remain unclosed when client abruptly disconnected (like closing a browser tab), but now it would close correctly
+- retrieve unserialized data from events with `ObjectProxy` (like JPEG images) by setting `deserialize=False` in `subscribe_event()` 
+- 
+
+## [v0.2.6] - 2024-09-09
+
+- bug fix events when multiple serializers are used
+- events support custom HTTP handlers (not polished yet, use as last resort, not auto-added to TD)
+- image event handlers for streaming live video as JPEG and PNG (not polished yet, not auto-added to TD)
+
+## [v0.2.5] - 2024-09-09
+
+- released to anaconda, it can take a while to turn up. A badge will be added in README when successful.  
 
 ## [v0.2.4] - 2024-09-09